project/blockchain/pyproject.toml
[project]
name = "blockchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "azure-functions>=1.23.0",
    "iota-sdk>=1.1.4",
    "psycopg2-binary>=2.9.10",
]
